neurowellness/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.7.4
pydantic-settings>=2.3.4
supabase==2.28.3
websockets>=13.0,<16
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
structlog==24.2.0
python-dotenv==1.0.1
slowapi==0.1.9
pytest==8.2.2
pytest-asyncio==0.23.7
ruff==0.4.10
